+++ lib/neocities_red.rb
@@ -5,13 +5,13 @@ require File.join(File.dirname(__FILE__), "neocities_red", "client")
 require File.join(File.dirname(__FILE__), "neocities_red", "cli_display")
 require File.join(File.dirname(__FILE__), "neocities_red", "cli")
 
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "file_uploader")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "folder_uploader")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "file_remover")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "file_list")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "file", "file")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "file", "folder_uploader")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "file", "remover")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "file", "list")
 
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "site_difference")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "profile_info")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "site_exporter")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "site_pusher")
-require File.join(File.dirname(__FILE__), "neocities_red", "services", "pizza")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "site", "differencer")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "site", "informer")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "site", "exporter")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "site", "pusher")
+require File.join(File.dirname(__FILE__), "neocities_red", "services", "common", "pizza")
