+++ lib/neocities_red/services/common/pizza.rb
@@ -1,0 +1,107 @@
+# frozen_string_literal: true
+
+require "pastel"
+
+module NeocitiesRed
+  module Services
+    module Common
+      class Pizza
+        EXCUSES = [
+          "Sorry, we're fresh out of pineapple today.",
+          "All the toppings just went rogue and are currently answering to no god.",
+          "Our bicycle delivery guy is out today for ska band practice.",
+          "Doughpocalypse now. Pizza's off until further notice.",
+          "Mamma mia! We're outta the cheesa.",
+          "The sauce of our youth ran dry. Pizza is off the menu for now.",
+          "There was this pizza place in Portland called Lonesomes that taped burned CDs of local bands to the pizza box. It was pretty dope.",
+          "No dough, no go, sorry joe.",
+          "I'll be right with you after I figure out how to center a div in CSS.",
+          "The pizza gods demand rest. Are the hunger pangs interrupting your game?",
+          "Our pizza chef currently has the high score on the Road Kings pinball machine, you dare disturb him?",
+          "Today's special: disappointment. Pizza unavailable.",
+          "Our last pizza became a perpetual motion machine, left the atmosphere and is flying through the heavens.",
+          "Ran out of oregano and optimism. See you next time.",
+          <<~TEXT
+            WAR AND PEACE, BY LEO TOLSTOY, BOOK ONE: 1805, CHAPTER I
+
+            “Well, Prince, so Genoa and Lucca are now just family estates of the Buonapartes.
+            But I warn you, if you don’t tell me that this means war, if you still try to defend
+            the infamies and horrors perpetrated by that Antichrist—I really believe he is
+            Antichrist—I will have nothing more to do with you and you are no longer my friend,
+            no longer my ‘faithful slave,’ as you call yourself! But how do you do? I see I have
+            frightened you—sit down and tell me all the news.”
+
+            It was in July, 1805, and the speaker was the well-known Anna Pávlovna Schérer,
+            maid of honor and favorite of the Empress Márya Fëdorovna. With these words she
+            greeted Prince Vasíli Kurágin, a man of high rank and importance, who was the first
+            to arrive at her reception.
+
+            Anna Pávlovna had had a cough for some days. She was, as she said, suffering from
+            la grippe; grippe being then a new word in St. Petersburg, used only by the elite.
+
+            All her invitations without exception, written in French, and delivered by a
+            scarlet-liveried footman that morning, ran as follows:
+
+            “If you have nothing better to do, Count (or Prince), and if the prospect of
+            spending an evening with a poor invalid is not too terrible, I shall be very
+            charmed to see you tonight between 7 and 10—Annette Schérer.”
+
+            “Heavens! what a virulent attack!” replied the prince, not in the least disconcerted
+            by this reception.
+
+            He had just entered, wearing an embroidered court uniform, knee breeches, and
+            shoes, and had stars on his breast and a serene expression on his flat face.
+
+            He spoke in that refined French in which our grandfathers not only spoke but
+            thought, and with the gentle, patronizing intonation natural to a man of importance
+            who had grown old in society and at court.
+
+            He went up to Anna Pávlovna, kissed her hand, presenting to her his bald, scented,
+            and shining head, and complacently seated himself on the sofa.
+
+            “First of all, dear friend, tell me how you are. Set your friend’s mind at rest,”
+            said he without altering his tone, beneath the politeness and affected sympathy of
+            which indifference and even irony could be discerned.
+
+            “Can one be well while suffering morally? Can one be calm in times like these if
+            one has any feeling?” said Anna Pávlovna. “You are staying the whole evening, I hope?”
+
+            “And the fete at the English ambassador’s? Today is Wednesday. I must put in an
+            appearance there,” said the prince. “My daughter is coming for me to take me there.”
+
+            “I thought today’s fete had been canceled. I confess all these festivities and
+            fireworks are becoming wearisome.”
+
+            “If they had known that you wished it, the entertainment would have been put off,”
+            said the prince, who, like a wound-up clock, by force of habit said things he did
+            not even wish to be believed.
+
+            “Don’t tease! Well, and what has been decided about Novosíltsev’s dispatch?
+            You know everything.”
+
+            “What can one say about it?” replied the prince in a cold, listless tone.
+            “What has been decided? They have decided that Buonaparte has burnt his boats,
+            and I believe that we are ready to burn ours.”
+
+            Prince Vasíli always spoke languidly, like an actor repeating a stale part.
+            Anna Pávlovna Schérer on the contrary, despite her forty years, overflowed with
+            animation and impulsiveness.
+
+            To be an enthusiast had become her social vocation and, sometimes even when she
+            did not feel like it, she became enthusiastic in order not to disappoint the
+            expectations of those who knew her.
+
+            The subdued smile which, though it did not suit her faded features, always played
+            round her lips expressed, as in a spoiled child, a continual consciousness of her
+            charming defect, which she neither wished, nor could, nor considered it necessary,
+            to correct.
+          TEXT
+        ].freeze
+
+        def make_order
+          Pastel.new(enabled: true).bright_red(EXCUSES.sample)
+        end
+      end
+    end
+  end
+end
