--- lib/neocities_red/services/file_uploader.rb
@@ -1,43 +1,0 @@
-# frozen_string_literal: true
-
-require "pathname"
-require "pastel"
-
-module NeocitiesRed
-  module Services
-    class FileIsNotExists < StandardError; end
-
-    class FileUploader
-      def initialize(client, filepath, remote_path = nil)
-        @client = client
-        @filepath = filepath
-        @remote_path = remote_path
-        @pastel = Pastel.new(eachline: "\n")
-      end
-
-      def upload
-        path = Pathname.new(@filepath)
-
-        raise FileIsNotExists, "#{path} does not exist locally." unless path.exist?
-
-        if path.directory?
-          puts @pastel.bold("#{path} is a directory, skipping")
-          return
-        end
-
-        puts @pastel.bold("Uploading #{path} to #{@remote_path} ...")
-
-        response = @client.upload(path, @remote_path)
-        puts response if response[:result] == "error"
-
-        if response[:result] == "error" && response[:error_type] == "file_exists"
-          puts @pastel.yellow.bold("EXISTS")
-        elsif response[:result] == "success"
-          puts @pastel.green.bold("SUCCESS")
-        end
-
-        response
-      end
-    end
-  end
-end
